fix(SavedMovies): card deleteHandler fixed
